Refactored CourseDetails Fragment
The Fragment Layout is now split in two parts.
The first part is the listview (now reusing CouresFragment for that)
The second part are the course details. These are now added as the header to the list view.
Since there is now only one listview and its header, the listview itself can handle all scrolling (since the details are part of the list
view)
This fixes #10